backend/requirements.txt
# Core LLM and Agent Framework
langgraph>=0.2.0
anthropic>=0.40.0
tavily-python>=0.5.0

# Vector Store and Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# PDF Processing
PyPDF2>=3.0.0

# Configuration Management
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Web Scraping (for Scout Agent)
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Optional: Progress tracking and logging
tqdm>=4.66.0
loguru>=0.7.0

# Development and Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.6.0
